+++ src/error.rs
@@ -1,3 +1,4 @@
+use crate::error::AllocError::ArithmeticOverflow;
 use alloc::alloc::Layout;
 use core::{
     fmt::{Debug, Display, Formatter, Result as FmtResult},
@@ -9,6 +10,11 @@ use core::{
 #[repr(u8)]
 #[allow(clippy::module_name_repetitions)]
 pub enum AllocError {
+    /// The underlying allocator failed to allocate using the given layout.
+    AllocFailed(Layout),
+    // commented out until implemented
+    // /// The underlying allocator failed to allocate using the given layout, with extra context.
+    // AllocFailedWithContext(Layout, Context),
     /// The layout computed with the given size and alignment is invalid.
     InvalidLayout(usize, usize),
     /// The given layout was zero-sized. The contained [`NonNull`] will be dangling and valid for
@@ -16,14 +22,13 @@ pub enum AllocError {
     ///
     /// This can, in many cases, be considered a success.
     ZeroSizedLayout(NonNull<u8>),
-    /// The underlying allocator failed to allocate using the given layout.
-    AllocFailed(Layout),
     /// Attempted to grow to a smaller layout.
     GrowSmallerNewLayout(usize, usize),
     /// Attempted to shrink to a larger layout.
     ShrinkBiggerNewLayout(usize, usize),
-    /// An arithmetic operation overflowed. This error contains the left-hand and right-hand side
-    /// values as well as the operation.
+    /// An arithmetic operation overflowed.
+    ///
+    /// This error contains both sides of the operation and the operation itself.
     ArithmeticOverflow(usize, ArithOp, usize),
     /// Any other kind of error, in the form of a string.
     Other(&'static str),
@@ -39,12 +44,19 @@ impl PartialEq for AllocError {
         };
 
         match (self, other) {
+            (AllocFailed(l1), AllocFailed(l2)) => l1 == l2,
+            // (
+            //     AllocError::AllocFailedWithContext(l1, c1),
+            //     AllocError::AllocFailedWithContext(l2, c2),
+            // ) => l1 == l2 && c1 == c2,
             (InvalidLayout(sz1, aln1), InvalidLayout(sz2, aln2)) => sz1 == sz2 && aln1 == aln2,
             (ZeroSizedLayout(a), ZeroSizedLayout(b)) => a == b,
-            (AllocFailed(l1), AllocFailed(l2)) => l1 == l2,
             (GrowSmallerNewLayout(old1, new1), GrowSmallerNewLayout(old2, new2))
             | (ShrinkBiggerNewLayout(old1, new1), ShrinkBiggerNewLayout(old2, new2)) => {
                 old1 == old2 && new1 == new2
+            }
+            (ArithmeticOverflow(lhs1, op1, rhs1), ArithmeticOverflow(lhs2, op2, rhs2)) => {
+                lhs1 == lhs2 && op1 == op2 && rhs1 == rhs2
             }
             (Other(a), Other(b)) => a == b,
             _ => false,
@@ -57,13 +69,16 @@ impl Eq for AllocError {}
 impl Display for AllocError {
     fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
         match self {
+            AllocError::AllocFailed(l) => write!(f, "allocation failed for layout: {:?}", l),
+            // AllocError::AllocFailedWithContext(l, c) => {
+            //     write!(f, "allocation failed for layout: {:?}, {}", l, c)
+            // }
             AllocError::InvalidLayout(sz, align) => {
                 write!(f, "computed invalid layout: size: {}, align: {}", sz, align)
             }
             AllocError::ZeroSizedLayout(_) => {
                 write!(f, "zero-sized layout was given")
             }
-            AllocError::AllocFailed(l) => write!(f, "allocation failed for layout: {:?}", l),
             AllocError::GrowSmallerNewLayout(old, new) => write!(
                 f,
                 "attempted to grow from a size of {} to a smaller size of {}",
@@ -111,3 +126,33 @@ impl Display for ArithOp {
         }
     }
 }
+
+// #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+// /// Context for an [`AllocError::AllocFailedWithContext`] error.
+// pub enum Context {
+//     /// The reason for the error.
+//     Reason(ErrorCause),
+// }
+//
+// impl Display for Context {
+//     fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+//         match self {
+//             Context::Reason(cause) => write!(f, "reason: {}", cause),
+//         }
+//     }
+// }
+//
+// #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+// /// The cause of an [`AllocError::AllocFailedWithContext`] error.
+// pub enum ErrorCause {
+//     /// The error was caused by running out of memory.
+//     OOM,
+// }
+//
+// impl Display for ErrorCause {
+//     fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+//         match self {
+//             ErrorCause::OOM => write!(f, "out of memory"),
+//         }
+//     }
+// }