+++ src/external_alloc/jemalloc.rs
@@ -25,40 +25,53 @@ macro_rules! assume {
 #[derive(Copy, Clone, Default, Debug, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub struct Jemalloc;
 
+fn alloc(layout: Layout) -> *mut c_void {
+    let flags = ffi::layout_to_flags(layout.size(), layout.align());
+    if flags == 0 {
+        unsafe { ffi::malloc(layout.size()) }
+    } else {
+        unsafe { ffi::mallocx(layout.size(), flags) }
+    }
+}
+
+fn alloc_zeroed(layout: Layout) -> *mut c_void {
+    let flags = ffi::layout_to_flags(layout.size(), layout.align());
+    if flags == 0 {
+        unsafe { ffi::calloc(1, layout.size()) }
+    } else {
+        unsafe { ffi::mallocx(layout.size(), flags | ffi::MALLOCX_ZERO) }
+    }
+}
+
+unsafe fn dealloc(ptr: *mut u8, layout: Layout) {
+    ffi::sdallocx(
+        ptr.cast::<c_void>(),
+        layout.size(),
+        ffi::layout_to_flags(layout.size(), layout.align()),
+    );
+}
+
 unsafe impl GlobalAlloc for Jemalloc {
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
         assume!(layout.size() != 0);
-        let flags = ffi::layout_to_flags(layout.size(), layout.align());
-        (if flags == 0 {
-            ffi::malloc(layout.size())
-        } else {
-            ffi::mallocx(layout.size(), flags)
-        }).cast::<u8>()
+        alloc(layout).cast::<u8>()
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn dealloc(&self, ptr: *mut u8, layout: Layout) {
         assume!(!ptr.is_null());
         assume!(layout.size() != 0);
-        ffi::sdallocx(
-            ptr.cast::<c_void>(),
-            layout.size(),
-            ffi::layout_to_flags(layout.size(), layout.align()),
-        );
+        dealloc(ptr, layout);
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn alloc_zeroed(&self, layout: Layout) -> *mut u8 {
         assume!(layout.size() != 0);
-        let flags = ffi::layout_to_flags(layout.size(), layout.align());
-        (if flags == 0 {
-            ffi::calloc(1, layout.size())
-        } else {
-            ffi::mallocx(layout.size(), flags | ffi::MALLOCX_ZERO)
-        }).cast::<u8>()
+        alloc_zeroed(layout).cast::<u8>()
     }
 
+    #[inline]
     unsafe fn realloc(&self, ptr: *mut u8, layout: Layout, new_size: usize) -> *mut u8 {
         assume!(layout.size() != 0);
         assume!(new_size != 0);
@@ -73,42 +86,27 @@ unsafe impl GlobalAlloc for Jemalloc {
 }
 
 impl Alloc for Jemalloc {
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     fn alloc(&self, layout: Layout) -> Result<NonNull<u8>, AllocError> {
         null_q_zsl_check(layout, |layout: Layout| {
-            let flags = ffi::layout_to_flags(layout.size(), layout.align());
-            if flags == 0 {
-                unsafe { ffi::malloc(layout.size()) }
-            } else {
-                unsafe { ffi::mallocx(layout.size(), flags) }
-            }
+            alloc(layout)
         })
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     fn alloc_zeroed(&self, layout: Layout) -> Result<NonNull<u8>, AllocError> {
         null_q_zsl_check(layout, |layout: Layout| {
-            let flags = ffi::layout_to_flags(layout.size(), layout.align());
-            if flags == 0 {
-                unsafe { ffi::calloc(1, layout.size()) }
-            } else {
-                unsafe { ffi::mallocx(layout.size(), flags | ffi::MALLOCX_ZERO) }
-            }
+            alloc_zeroed(layout)
         })
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn dealloc(&self, ptr: NonNull<u8>, layout: Layout) {
         if layout.size() != 0 {
-            ffi::sdallocx(
-                ptr.as_ptr().cast::<c_void>(),
-                layout.size(),
-                ffi::layout_to_flags(layout.size(), layout.align()),
-            );
+            dealloc(ptr.as_ptr(), layout);
         }
     }
 
-    #[cfg_attr(miri, track_caller)]
     unsafe fn grow(
         &self,
         ptr: NonNull<u8>,
@@ -125,7 +123,6 @@ impl Alloc for Jemalloc {
         )
     }
 
-    #[cfg_attr(miri, track_caller)]
     unsafe fn shrink(
         &self,
         ptr: NonNull<u8>,
@@ -158,7 +155,6 @@ impl Alloc for Jemalloc {
     ///
     /// - `ptr` must point to a block previously allocated with this allocator.
     /// - `old_layout` must describe exactly that block.
-    #[cfg_attr(miri, track_caller)]
     unsafe fn realloc(
         &self,
         ptr: NonNull<u8>,