+++ src/external_alloc/mimalloc.rs
@@ -13,28 +13,28 @@ use libc::c_void;
 pub struct MiMalloc;
 
 unsafe impl GlobalAlloc for MiMalloc {
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
         ffi::mi_malloc_aligned(layout.size(), layout.align()).cast::<u8>()
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn dealloc(&self, ptr: *mut u8, layout: Layout) {
         ffi::mi_free_size_aligned(ptr.cast::<c_void>(), layout.size(), layout.align());
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn alloc_zeroed(&self, layout: Layout) -> *mut u8 {
         ffi::mi_zalloc_aligned(layout.size(), layout.align()).cast::<u8>()
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn realloc(&self, ptr: *mut u8, layout: Layout, new_size: usize) -> *mut u8 {
         ffi::mi_realloc_aligned(ptr.cast::<c_void>(), new_size, layout.align()).cast::<u8>()
     }
 }
 
-#[cfg_attr(miri, track_caller)]
+#[inline]
 fn zsl_check_alloc(
     layout: Layout,
     alloc: unsafe extern "C" fn(usize, usize) -> *mut c_void,
@@ -45,24 +45,23 @@ fn zsl_check_alloc(
 }
 
 impl Alloc for MiMalloc {
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     fn alloc(&self, layout: Layout) -> Result<NonNull<u8>, AllocError> {
         zsl_check_alloc(layout, ffi::mi_malloc_aligned)
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     fn alloc_zeroed(&self, layout: Layout) -> Result<NonNull<u8>, AllocError> {
         zsl_check_alloc(layout, ffi::mi_zalloc_aligned)
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn dealloc(&self, ptr: NonNull<u8>, layout: Layout) {
         if layout.size() != 0 {
             ffi::mi_free_size_aligned(ptr.as_ptr().cast::<c_void>(), layout.size(), layout.align());
         }
     }
 
-    #[cfg_attr(miri, track_caller)]
     unsafe fn grow(
         &self,
         ptr: NonNull<u8>,
@@ -85,7 +84,6 @@ impl Alloc for MiMalloc {
         )
     }
 
-    #[cfg_attr(miri, track_caller)]
     unsafe fn shrink(
         &self,
         ptr: NonNull<u8>,
@@ -108,7 +106,7 @@ impl Alloc for MiMalloc {
         )
     }
 
-    #[cfg_attr(miri, track_caller)]
+    #[inline]
     unsafe fn realloc(
         &self,
         ptr: NonNull<u8>,